add dir to runner and make gitService a member restored the isTaggable calculation to the parser
